refactor: chatRoomViewController에 있는 비즈니스 로직을 UseCase로 분리
